requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings
celery==5.3.4
redis==5.0.1
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
opencv-python==4.8.1.78
numpy==1.24.3
pillow==10.1.0
scipy==1.11.4
scikit-image==0.22.0
prometheus-client==0.19.0
prometheus-fastapi-instrumentator==6.1.0
rich==13.7.0
python-multipart==0.0.6
aiofiles==23.2.1
httpx==0.25.2
python-dotenv==1.0.0